Move the definition of a row in its own file
Move the definition of the columns in the same file as a compromise
