+++ Source/PlasticSourceControl/Private/SPlasticSourceControlBranchesWidget.h
@@ -5,16 +5,12 @@
 #include "CoreMinimal.h"
 #include "Misc/TextFilter.h"
 #include "Widgets/SCompoundWidget.h"
-#include "Widgets/Views/SHeaderRow.h"
 #include "Widgets/Views/SListView.h"
-#include "Widgets/Views/STableRow.h"
-#include "Widgets/Views/STableViewBase.h"
 
 #include "ISourceControlOperation.h"
 #include "ISourceControlProvider.h"
 
 typedef TSharedRef<class FPlasticSourceControlBranch, ESPMode::ThreadSafe> FPlasticSourceControlBranchRef;
-typedef TSharedPtr<class FPlasticSourceControlBranch, ESPMode::ThreadSafe> FPlasticSourceControlBranchPtr;
 
 class SSearchBox;
 
@@ -87,90 +83,4 @@ private:
 
 	TArray<FPlasticSourceControlBranchRef> SourceControlBranches; // Full list from source (filtered by date)
 	TArray<FPlasticSourceControlBranchRef> BranchRows; // Filtered list to display based on the search text filter
-};
-
-
-// TODO: move the following to its own file
-// Inspired by class SFileTableRow : public SMultiColumnTableRow<FChangelistTreeItemPtr> in SSourceControlChangelistRows.h
-
-
-/** Lists the unique columns used in the list view displaying branches. */
-namespace PlasticSourceControlBranchesListViewColumn
-{
-	/** The branch Name column. */
-	namespace Name // NOLINT(runtime/indentation_namespace)
-	{
-		FName Id();
-		FText GetDisplayText();
-		FText GetToolTipText();
-	};
-
-	/** The branch Repository column. */
-	namespace Repository // NOLINT(runtime/indentation_namespace)
-	{
-		FName Id();
-		FText GetDisplayText();
-		FText GetToolTipText();
-	};
-
-	/** The branch CreatedBy column. */
-	namespace CreatedBy // NOLINT(runtime/indentation_namespace)
-	{
-		FName Id();
-		FText GetDisplayText();
-		FText GetToolTipText();
-	};
-
-	/** The branch Date column. */
-	namespace Date // NOLINT(runtime/indentation_namespace)
-	{
-		FName Id();
-		FText GetDisplayText();
-		FText GetToolTipText();
-	};
-
-	/** The branch Comment column. */
-	namespace Comment // NOLINT(runtime/indentation_namespace)
-	{
-		FName Id();
-		FText GetDisplayText();
-		FText GetToolTipText();
-	}
-} // namespace PlasticSourceControlBranchesListViewColumn
-
-class SBranchTableRow : public SMultiColumnTableRow<FPlasticSourceControlBranchRef>
-{
-public:
-	SLATE_BEGIN_ARGS(SBranchTableRow)
-		: _BranchToVisualize(nullptr)
-		, _bIsCurrentBranch(false)
-		, _HighlightText()
-	{
-	}
-		SLATE_ARGUMENT(FPlasticSourceControlBranchPtr, BranchToVisualize)
-		SLATE_ATTRIBUTE(bool, bIsCurrentBranch)
-		SLATE_ATTRIBUTE(FText, HighlightText)
-	SLATE_END_ARGS()
-
-public:
-	/**
-	* Construct a row child widgets of the ListView.
-	*
-	* @param InArgs Parameters including the branch to visualize in this row.
-	* @param InOwner The owning ListView.
-	*/
-	void Construct(const FArguments& InArgs, const TSharedRef<STableViewBase>& InOwner);
-
-	// SMultiColumnTableRow overrides
-	virtual TSharedRef<SWidget> GenerateWidgetForColumn(const FName& ColumnId) override;
-
-private:
-	/** The branch that we are visualizing in this row. */
-	FPlasticSourceControlBranch* BranchToVisualize;
-
-	/** True if this is the current branch, to be highlighted on the list of branches. */
-	bool bIsCurrentBranch;
-
-	/** The search text to highlight if any */
-	TAttribute<FText> HighlightText;
 };
